% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/arrange_tows.R
\name{arrange_tows}
\alias{arrange_tows}
\title{Arrange the tows so they all are in the same direction
Function to Orient all tows in the same direction}
\usage{
arrange_tows(dat, direction = "NW")
}
\arguments{
\item{dat}{Input data}

\item{direction}{Direction to orient all tows in. Default is "NW" and everything is hard coded 
which I could change at some point}
}
\description{
Arrange the tows so they all are in the same direction
Function to Orient all tows in the same direction
}
\examples{
arrange_tows(dat = part_wc_data, direction = 'NW')
}

