% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_deltas_spp.R
\name{calc_deltas_spp}
\alias{calc_deltas_spp}
\title{Calculate Delta Species
Function to Calculate delta plot values given a vector of species. This function is called by
delta_plot and applied across a vector of years}
\usage{
calc_deltas_spp(data, spps, focus = "hpounds")
}
\arguments{
\item{data}{Data that has catches per tow}

\item{spps}{Vector of species of interest}

\item{focus}{Column of focus, defaults to hpounds}
}
\description{
Calculate Delta Species
Function to Calculate delta plot values given a vector of species. This function is called by
delta_plot and applied across a vector of years
}
\examples{
calc_deltas_spp(data = wc_data,)
}

