% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_skew.R
\name{calc_skew}
\alias{calc_skew}
\title{Calculate skew of a catch distribution
Function to calculate the skew of a distribution (third moment)}
\usage{
calc_skew(input)
}
\arguments{
\item{input}{Vector of values}
}
\description{
Calculate skew of a catch distribution
Function to calculate the skew of a distribution (third moment)
}
\examples{
calc_skew(input = rnorm(1000))
}

