% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_tows.R
\name{clust_tows}
\alias{clust_tows}
\title{Cluster the data
Funtion that returns cluster output. The purpose is to allow clustering to occur only once
while exploring effects of different cut points}
\usage{
clust_tows(dat)
}
\arguments{
\item{dat}{Input data}
}
\description{
Cluster the data
Funtion that returns cluster output. The purpose is to allow clustering to occur only once
while exploring effects of different cut points
}
\examples{
clust_tows(dat = dat)
}

