% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_two_spp.R
\name{compare_two_spp}
\alias{compare_two_spp}
\title{Compare metrics between two species
Function to count number of tows in bins. Uses ggplot stat_bin2d functions.}
\usage{
compare_two_spp(data = wc_data, species1, species2, focus = "hperc")
}
\arguments{
\item{data}{Defaults to wc_data}

\item{species1}{Species 1 to look at}

\item{species2}{Species 2 to look at}

\item{focus}{Column of interest, Options are hperc, aperc, hpounds, apounds}
}
\description{
Compare metrics between two species
Function to count number of tows in bins. Uses ggplot stat_bin2d functions.
}
\examples{
compare_two_spp(data = wc_data, species1 = "Dover Sole", species2 = "Sablefish", focus = 'hperc')
}

