% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cut_for_merge.R
\name{cut_for_merge}
\alias{cut_for_merge}
\title{Cut for merge
Cut cluster data for merge with original data frame. Output only has haul_id, clust,
and number of tows}
\usage{
cut_for_merge(input_data, clust_input, cut_point)
}
\arguments{
\item{input_data}{Original data input to function. Should be only the unique tows}

\item{clust_input}{Output from the clustering function. Clustering can take a while}

\item{cut_point}{Value to cut the data}
}
\description{
Cut for merge
Cut cluster data for merge with original data frame. Output only has haul_id, clust,
and number of tows
}
\examples{
cut_for_merge(input_data = unq_tows, clust_input = xx, cut_point = calc_cut)
}

