% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plot.R
\name{delta_plot}
\alias{delta_plot}
\title{Create delta plots
Function to generate values relevant for delta plots. Delta plots come from Gillis et al. 
(2008)}
\usage{
delta_plot()
}
\arguments{
\item{none}{There are no parameters}
}
\description{
Create delta plots
Function to generate values relevant for delta plots. Delta plots come from Gillis et al. 
(2008)
}
\examples{
delta_plot()
}

