% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_data.R
\name{filter_data}
\alias{filter_data}
\title{Filter Data Over Polygons
Function to filter data points over shape file}
\usage{
filter_data(input, shape_dir = "all_data/wc_shapefile",
  file_nm = "Strata_Final_dissAreaType",
  proj = "+proj=longlat +datum=WGS84 +ellps=WGS84 +towgs84=0,0,0")
}
\arguments{
\item{input}{Points to filter, preferably subsetted by speed}

\item{shape_dir}{Directory of shapefile}

\item{file_nm}{Shapefile name}

\item{proj}{Projection characteristics}
}
\description{
Filter Data Over Polygons
Function to filter data points over shape file
}
