% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_ne_vms.R
\name{load_ne_vms}
\alias{load_ne_vms}
\title{Load Northeast VMS Data
Function to load northeast VMS Data}
\usage{
load_ne_vms()
}
\arguments{
\item{Parameter}{there aren't any}
}
\description{
Load Northeast VMS Data
Function to load northeast VMS Data
}
\examples{
load_ne_vms()
}

