% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wc_itq.R
\name{load_wc_itq}
\alias{load_wc_itq}
\title{Function to load west coast VMS data and subset for bottom trawl only}
\usage{
load_wc_itq()
}
\arguments{
\item{none}{No parameters}
}
\description{
Function to load west coast VMS data and subset for bottom trawl only
}
\examples{
load_wc_itq()
}
