% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_wc_logbook.R
\name{load_wc_logbook}
\alias{load_wc_logbook}
\title{Load West Coast Logbook Data
Function to load West Coast Logbook data}
\usage{
load_wc_logbook()
}
\arguments{
\item{Parameter}{there aren't any}
}
\description{
Load West Coast Logbook Data
Function to load West Coast Logbook data
}
\examples{
load_wc_logbook()
}

