% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hlsimulator.R
\name{multiplot}
\alias{multiplot}
\title{Multiplot function
Function to make panel plots in ggplot. Copied this from 
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/}
\usage{
multiplot(..., plotlist = NULL, file, cols = 1, layout = NULL)
}
\arguments{
\item{...}{ggplot inputs}

\item{plotlist}{don't know}

\item{file}{not sure}

\item{cols}{Number of columns}

\item{file}{not sure}
}
\description{
Multiplot function
Function to make panel plots in ggplot. Copied this from 
http://www.cookbook-r.com/Graphs/Multiple_graphs_on_one_page_(ggplot2)/
}
