% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/na_to_zero.R
\name{na_to_zero}
\alias{na_to_zero}
\title{Convert missing values to zeroes
Function to convert NAs to 0}
\usage{
na_to_zero(vector_in)
}
\arguments{
\item{vector_in}{Vector to manipulate}
}
\description{
Convert missing values to zeroes
Function to convert NAs to 0
}
\examples{
na_to_zero(c(NA, NA, 2))
}

