% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pull_catch.R
\name{pull_catch}
\alias{pull_catch}
\title{Find catch from unique tows
Function to count number of tows in bins. Uses ggplot stat_bin2d functions.}
\usage{
pull_catch(data = wc_data, spp, focus = "apound")
}
\arguments{
\item{data}{Input data, defaults to wc_data}

\item{spp}{Species of interest}

\item{focus}{Column of interest, Options are hpounds, apounds}
}
\description{
Find catch from unique tows
Function to count number of tows in bins. Uses ggplot stat_bin2d functions.
}
\examples{
pull_catch(data = wc_data, spp = 'Dover Sole', focus = 'apound')
compare_two_spp(data = wc_data, species1 = "Dover Sole", species2 = "Sablefish", focus = 'hperc')
}

