% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resample_years.R
\name{resample_years}
\alias{resample_years}
\title{Resample years of logbook data
Function to permute catch data and evaluate significant declines/increases in clusters.}
\usage{
resample_years(d_in, niters, seed = 1)
}
\arguments{
\item{d_in}{Data input into the function}

\item{niters}{Number of iterations for each resampling}

\item{seed}{Seed value; defaults to 1}
}
\description{
Resample years of logbook data
Function to permute catch data and evaluate significant declines/increases in clusters.
}
\examples{
resample_years(d_in = data, niters = 10)
}

