% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_clust.R
\name{summ_clust}
\alias{summ_clust}
\title{Summarize Clusters
Function to summarize clusters, given a certain cut_point}
\usage{
summ_clust(dat, clust_input, cut_point)
}
\arguments{
\item{dat}{Original data that was clustered}

\item{clust_input}{Output from clust_tows function}

\item{cut_point}{Cut point}
}
\description{
Summarize Clusters
Function to summarize clusters, given a certain cut_point
}
\examples{
Coming soon
}

