% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clust_tows.R
\name{clust_tows}
\alias{clust_tows}
\title{Cluster the data
Funtion that returns cluster output. The purpose is to allow clustering to occur only once
while exploring effects of different cut points}
\usage{
clust_tows(dat, port)
}
\arguments{
\item{dat}{Input data}

\item{port}{Port of interest. Function takes too long if done on all data}
}
\description{
Cluster the data
Funtion that returns cluster output. The purpose is to allow clustering to occur only once
while exploring effects of different cut points
}
\examples{
clust_tows(dat = dat, port = 'ASTORIA')
}

