% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_plot.R
\name{delta_plot}
\alias{delta_plot}
\title{Create delta plots
Function to generate values relevant for delta plots. Delta plots come from Gillis et al. 
(2008)}
\usage{
delta_plot(data, spps = tolower(c("Dover Sole", "Arrowtooth Flounder",
  "Sablefish", "Petrale Sole", "Longspine Thornyhead", "Shortspine Thornyhead",
  "Chilipepper Rockfish", "Lingcod", "Yellowtail Rockfish",
  "Darkblotched Rockfish", "Pacific Ocean Perch", "Bank Rockfish",
  "Widow Rockfish")), year_col = "tow_year")
}
\arguments{
\item{data}{Data frame used to calculate delta plots}

\item{spps}{Vector of species to calculate delta plot values for}

\item{year_col}{Name of column containing year information, defaults to tow_year which
corresponds to wc_data}
}
\description{
Create delta plots
Function to generate values relevant for delta plots. Delta plots come from Gillis et al. 
(2008)
}
\examples{
delta_plot()
}

