% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summ_clust.R
\name{summ_clust}
\alias{summ_clust}
\title{Summarize Clusters
Function to summarize clusters, given a certain cut_point}
\usage{
summ_clust(dat)
}
\arguments{
\item{dat}{Data with column for clusters}
}
\description{
Summarize Clusters
Function to summarize clusters, given a certain cut_point
}
\examples{
Coming soon
}

